requirements.txt
pyvcf==0.6.8